fix: a couple of fixes to the example
