+++ examples/minimal-convex/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@example/minimal-appdir",
+  "name": "@example/minimal-convex",
   "private": true,
   "scripts": {
     "dev": "next dev",
