+++ examples/minimal-convex/tsconfig.json
@@ -11,6 +11,7 @@
     "jsx": "preserve",
     "incremental": true,
     "plugins": [{ "name": "next" }],
+    "baseUrl": ".",
     "paths": {
       "~/*": ["./src/*"],
       "~/convex/*": ["./convex/*"]
